backend/requirements.txt
alpha_vantage==2.3.1
yfinance==0.2.50
pycoingecko==3.1.0
django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
psycopg2-binary==2.9.6 
python-binance==1.0.22
requests
beautifulsoup4
